Move more color-cycling logic into the color-cycling module.

+++ src/render/painter.cpp
@@ -143,8 +143,7 @@ void Painter::add_mods( VertexBase&        vert,
     vert.set_uniform_depixelation( *mods.uniform_depixelation );
   vert.set_color_cycle( mods.cycling.plan.has_value() );
   if( mods.cycling.plan.has_value() )
-    vert.set_aux_idx(
-        static_cast<int32_t>( *mods.cycling.plan ) );
+    vert.set_aux_idx( *mods.cycling.plan );
 }
 
 Painter& Painter::draw_point( point p, pixel color ) {
